+++ src/main/java/team/voided/quiltenergy/energy/interaction/EnergyInteractionResult.java
@@ -1,5 +1,6 @@
 package team.voided.quiltenergy.energy.interaction;
 
 import team.voided.quiltenergy.energy.EnergyUnit;
+import team.voided.quiltenergy.numerics.Decimal;
 
-public record EnergyInteractionResult(EnergyUnit unit, double originalAmount, double newAmount, boolean succeeded) { }
+public record EnergyInteractionResult(EnergyUnit unit, Decimal originalAmount, Decimal newAmount, boolean succeeded) { }
